requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.0
django-timezone-field==7.1
djangorestframework==3.16.0
frozenlist==1.6.0
idna==3.10
kombu==5.5.3
magic-filter==1.0.12
multidict==6.4.3
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13c
yarl==1.20.0
